+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 4.0.1 under development
 
-- no changes in this release.
+- Bug #142: Fix dark mode argument display issues (@pamparam83)
 
 ## 4.0.0 February 05, 2025